+++ azext_iot/_help.py
@@ -298,38 +298,70 @@ helps[
     examples:
     - name: Get the parent device of the specified device.
       text: >
-        az iot hub device-identity get-parent -d {non_edge_device_id} -n {iothub_name}
+        az iot hub device-identity get-parent -d {device_id} -n {iothub_name}
 """
 
 helps[
     "iot hub device-identity set-parent"
 ] = """
     type: command
-    short-summary: Set the parent device of the specified non-edge device.
+    short-summary: Set the parent device of the specified device.
     examples:
-    - name: Set the parent device of the specified non-edge device.
+    - name: Set the parent device of the specified device.
       text: >
-        az iot hub device-identity set-parent -d {non_edge_device_id} --pd {edge_device_id} -n {iothub_name}
-    - name: Set the parent device of the specified non-edge device irrespectively the non-edge device is
+        az iot hub device-identity set-parent -d {device_id} --pd {edge_device_id} -n {iothub_name}
+    - name: Set the parent device of the specified device irrespectively the device is
             already a child of other edge device.
       text: >
-        az iot hub device-identity set-parent -d {non_edge_device_id} --pd {edge_device_id} --force -n {iothub_name}
+        az iot hub device-identity set-parent -d {device_id} --pd {edge_device_id} --force -n {iothub_name}
+"""
+
+helps[
+    "iot hub device-identity parent"
+] = """
+    type: group
+    short-summary: Manage IoT device\'s parent device.
+"""
+
+helps[
+    "iot hub device-identity parent show"
+] = """
+    type: command
+    short-summary: Get the parent device of the specified device.
+    examples:
+    - name: Get the parent device of the specified device.
+      text: >
+        az iot hub device-identity parent show -d {device_id} -n {iothub_name}
+"""
+
+helps[
+    "iot hub device-identity parent set"
+] = """
+    type: command
+    short-summary: Set the parent device of the specified device.
+    examples:
+    - name: Set the parent device of the specified device.
+      text: >
+        az iot hub device-identity parent set -d {device_id} --pd {edge_device_id} -n {iothub_name}
+    - name: Set the parent device of the specified device and overwrites its original parent.
+      text: >
+        az iot hub device-identity parent set -d {device_id} --pd {edge_device_id} --force -n {iothub_name}
 """
 
 helps[
     "iot hub device-identity add-children"
 ] = """
     type: command
-    short-summary: Add specified comma-separated list of non edge device ids as children of specified edge device.
+    short-summary: Add specified comma-separated list of device ids as children of specified edge device.
     examples:
-    - name: Add non-edge devices as a children to the edge device.
+    - name: Add devices as a children to the edge device.
       text: >
-        az iot hub device-identity add-children -d {edge_device_id} --child-list {comma_separated_non_edge_device_id}
+        az iot hub device-identity add-children -d {edge_device_id} --child-list {comma_separated_device_id}
         -n {iothub_name}
-    - name: Add non-edge devices as a children to the edge device irrespectively the non-edge device is
+    - name: Add devices as a children to the edge device irrespectively the device is
             already a child of other edge device.
       text: >
-        az iot hub device-identity add-children -d {edge_device_id} --child-list {comma_separated_non_edge_device_id}
+        az iot hub device-identity add-children -d {edge_device_id} --child-list {comma_separated_device_id}
         -n {iothub_name} -f
 """
 
@@ -337,9 +369,9 @@ helps[
     "iot hub device-identity list-children"
 ] = """
     type: command
-    short-summary: Print comma-separated list of assigned child devices.
+    short-summary: Outputs comma-separated list of assigned child devices.
     examples:
-    - name: Show all assigned non-edge devices as comma-separated list.
+    - name: Show all assigned devices as comma-separated list.
       text: >
         az iot hub device-identity list-children -d {edge_device_id} -n {iothub_name}
 """
@@ -348,15 +380,68 @@ helps[
     "iot hub device-identity remove-children"
 ] = """
     type: command
-    short-summary: Remove non edge devices as children from specified edge device.
+    short-summary: Remove devices as children from specified edge device.
     examples:
     - name: Remove all mentioned devices as children of specified device.
       text: >
-        az iot hub device-identity remove-children -d {edge_device_id} --child-list {comma_separated_non_edge_device_id}
+        az iot hub device-identity remove-children -d {edge_device_id} --child-list {comma_separated_device_id}
         -n {iothub_name}
-    - name: Remove all non-edge devices as children specified edge device.
+    - name: Remove all devices as children specified edge device.
       text: >
         az iot hub device-identity remove-children -d {edge_device_id} --remove-all
+"""
+
+helps[
+    "iot hub device-identity children"
+] = """
+    type: group
+    short-summary: Manage IoT device\'s children device.
+"""
+
+helps[
+    "iot hub device-identity children add"
+] = """
+    type: command
+    short-summary: Add specified space-separated list of device ids as children of specified edge device.
+    examples:
+    - name: Add devices as a children to the edge device.
+      text: >
+        az iot hub device-identity children add -d {edge_device_id} --child-list {space_separated_device_id}
+        -n {iothub_name}
+    - name: Add devices as children to the edge device and overwrites children devices'
+            original parent.
+      text: >
+        az iot hub device-identity children add -d {edge_device_id} --child-list {space_separated_device_id}
+        -n {iothub_name} -f
+"""
+
+helps[
+    "iot hub device-identity children list"
+] = """
+    type: command
+    short-summary: Outputs list of assigned child devices.
+    examples:
+    - name: Show all assigned children devices as list.
+      text: >
+        az iot hub device-identity children list -d {edge_device_id} -n {iothub_name}
+    - name: Show all assigned children devices as list whose device ID contains substring of 'test'.
+      text: >
+        az iot hub device-identity children list -d {edge_device_id} -n {iothub_name} --query "[?contains(@,'test')]"
+"""
+
+helps[
+    "iot hub device-identity children remove"
+] = """
+    type: command
+    short-summary: Remove devices as children from specified edge device.
+    examples:
+    - name: Remove all mentioned devices as children of specified device.
+      text: >
+        az iot hub device-identity children remove -d {edge_device_id} --child-list {space_separated_device_id}
+        -n {iothub_name}
+    - name: Remove all devices as children specified edge device.
+      text: >
+        az iot hub device-identity children remove -d {edge_device_id} --remove-all
 """
 
 helps[