+++ azext_iot/sdk/iothub/service/models/query_result_py3.py
@@ -21,7 +21,7 @@ class QueryResult(Model):
     :type type: str or ~service.models.enum
     :param items: The query result items, as a collection.
     :type items: list[object]
-    :param continuation_token: Request continuation token.
+    :param continuation_token: The continuation token.
     :type continuation_token: str
     """